Updates and cleanup
