+++ cs/src/core/Index/Recovery/LocalCheckpointManager.cs
@@ -29,17 +29,35 @@ namespace FASTER.core
         }
 
         /// <summary>
-        /// 
+        /// Initialize index checkpoint
         /// </summary>
         /// <param name="indexToken"></param>
-        /// <param name="commitInfo"></param>
-        public void CommitIndexCheckpoint(Guid indexToken, byte[] commitInfo)
+        public void InitializeIndexCheckpoint(Guid indexToken)
+        {
+            directoryConfiguration.CreateIndexCheckpointFolder(indexToken);
+        }
+
+        /// <summary>
+        /// Initialize log checkpoint (snapshot and fold-over)
+        /// </summary>
+        /// <param name="logToken"></param>
+        public void InitializeLogCheckpoint(Guid logToken)
+        {
+            directoryConfiguration.CreateHybridLogCheckpointFolder(logToken);
+        }
+
+        /// <summary>
+        /// Commit index checkpoint
+        /// </summary>
+        /// <param name="indexToken"></param>
+        /// <param name="commitMetadata"></param>
+        public void CommitIndexCheckpoint(Guid indexToken, byte[] commitMetadata)
         {
             string filename = directoryConfiguration.GetIndexCheckpointMetaFileName(indexToken);
             using (var writer = new BinaryWriter(new FileStream(filename, FileMode.Create)))
             {
-                writer.Write(commitInfo.Length);
-                writer.Write(commitInfo);
+                writer.Write(commitMetadata.Length);
+                writer.Write(commitMetadata);
                 writer.Flush();
             }
 
@@ -52,17 +70,17 @@ namespace FASTER.core
         }
 
         /// <summary>
-        /// 
+        /// Commit log checkpoint (snapshot and fold-over)
         /// </summary>
         /// <param name="logToken"></param>
-        /// <param name="commitInfo"></param>
-        public void CommitLogCheckpoint(Guid logToken, byte[] commitInfo)
+        /// <param name="commitMetadata"></param>
+        public void CommitLogCheckpoint(Guid logToken, byte[] commitMetadata)
         {
             string filename = directoryConfiguration.GetHybridLogCheckpointMetaFileName(logToken);
             using (var writer = new BinaryWriter(new FileStream(filename, FileMode.Create)))
             {
-                writer.Write(commitInfo.Length);
-                writer.Write(commitInfo);
+                writer.Write(commitMetadata.Length);
+                writer.Write(commitMetadata);
                 writer.Flush();
             }
 
@@ -75,10 +93,10 @@ namespace FASTER.core
         }
 
         /// <summary>
-        /// 
+        /// Retrieve commit metadata for specified index checkpoint
         /// </summary>
-        /// <param name="indexToken"></param>
-        /// <returns></returns>
+        /// <param name="indexToken">Token</param>
+        /// <returns>Metadata, or null if invalid</returns>
         public byte[] GetIndexCommitMetadata(Guid indexToken)
         {
             var dir = new DirectoryInfo(directoryConfiguration.GetIndexCheckpointFolder(indexToken));
@@ -94,7 +112,26 @@ namespace FASTER.core
         }
 
         /// <summary>
-        /// 
+        /// Retrieve commit metadata for specified log checkpoint
+        /// </summary>
+        /// <param name="logToken">Token</param>
+        /// <returns>Metadata, or null if invalid</returns>
+        public byte[] GetLogCommitMetadata(Guid logToken)
+        {
+            var dir = new DirectoryInfo(directoryConfiguration.GetHybridLogCheckpointFolder(logToken));
+            if (!File.Exists(dir.FullName + Path.DirectorySeparatorChar + "completed.dat"))
+                return null;
+
+            string checkpointInfoFile = directoryConfiguration.GetHybridLogCheckpointMetaFileName(logToken);
+            using (var reader = new BinaryReader(new FileStream(checkpointInfoFile, FileMode.Open)))
+            {
+                var len = reader.ReadInt32();
+                return reader.ReadBytes(len);
+            }
+        }
+
+        /// <summary>
+        /// Provide device to store index checkpoint (including overflow buckets)
         /// </summary>
         /// <param name="indexToken"></param>
         /// <returns></returns>
@@ -104,7 +141,27 @@ namespace FASTER.core
         }
 
         /// <summary>
-        /// 
+        /// Provide device to store snapshot of log (required only for snapshot checkpoints)
+        /// </summary>
+        /// <param name="token"></param>
+        /// <returns></returns>
+        public IDevice GetSnapshotLogDevice(Guid token)
+        {
+            return Devices.CreateLogDevice(directoryConfiguration.GetLogSnapshotFileName(token), false);
+        }
+
+        /// <summary>
+        /// Provide device to store snapshot of object log (required only for snapshot checkpoints)
+        /// </summary>
+        /// <param name="token"></param>
+        /// <returns></returns>
+        public IDevice GetSnapshotObjectLogDevice(Guid token)
+        {
+            return Devices.CreateLogDevice(directoryConfiguration.GetObjectLogSnapshotFileName(token), false);
+        }
+
+        /// <summary>
+        /// Get latest valid checkpoint for recovery
         /// </summary>
         /// <param name="indexToken"></param>
         /// <param name="logToken"></param>
@@ -144,190 +201,6 @@ namespace FASTER.core
                 throw new Exception("No valid hybrid log checkpoint to recover from");
             }
             return true;
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="logToken"></param>
-        /// <returns></returns>
-        public byte[] GetLogCommitMetadata(Guid logToken)
-        {
-            var dir = new DirectoryInfo(directoryConfiguration.GetHybridLogCheckpointFolder(logToken));
-            if (!File.Exists(dir.FullName + Path.DirectorySeparatorChar + "completed.dat"))
-                return null;
-
-            string checkpointInfoFile = directoryConfiguration.GetHybridLogCheckpointMetaFileName(logToken);
-            using (var reader = new BinaryReader(new FileStream(checkpointInfoFile, FileMode.Open)))
-            {
-                var len = reader.ReadInt32();
-                return reader.ReadBytes(len);
-            }
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="token"></param>
-        /// <returns></returns>
-        public IDevice GetSnapshotLogDevice(Guid token)
-        {
-            return Devices.CreateLogDevice(directoryConfiguration.GetLogSnapshotFileName(token), false);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="token"></param>
-        /// <returns></returns>
-        public IDevice GetSnapshotObjectLogDevice(Guid token)
-        {
-            return Devices.CreateLogDevice(directoryConfiguration.GetObjectLogSnapshotFileName(token), false);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="indexToken"></param>
-        public void InitializeIndexCheckpoint(Guid indexToken)
-        {
-            directoryConfiguration.CreateIndexCheckpointFolder(indexToken);
-        }
-
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="logToken"></param>
-        public void InitializeLogCheckpoint(Guid logToken)
-        {
-            directoryConfiguration.CreateHybridLogCheckpointFolder(logToken);
-        }
-    }
-
-
-    class DirectoryConfiguration
-    {
-        private readonly string checkpointDir;
-
-        public DirectoryConfiguration(string checkpointDir)
-        {
-            this.checkpointDir = checkpointDir;
-        }
-
-        public const string index_base_folder = "index-checkpoints";
-        public const string index_meta_file = "info";
-        public const string hash_table_file = "ht";
-        public const string overflow_buckets_file = "ofb";
-        public const string snapshot_file = "snapshot";
-
-        public const string cpr_base_folder = "cpr-checkpoints";
-        public const string cpr_meta_file = "info";
-
-        public void CreateIndexCheckpointFolder(Guid token)
-        {
-            var directory = GetIndexCheckpointFolder(token);
-            Directory.CreateDirectory(directory);
-            DirectoryInfo directoryInfo = new System.IO.DirectoryInfo(directory);
-            foreach (System.IO.FileInfo file in directoryInfo.GetFiles())
-                file.Delete();
-        }
-        public void CreateHybridLogCheckpointFolder(Guid token)
-        {
-            var directory = GetHybridLogCheckpointFolder(token);
-            Directory.CreateDirectory(directory);
-            DirectoryInfo directoryInfo = new System.IO.DirectoryInfo(directory);
-            foreach (System.IO.FileInfo file in directoryInfo.GetFiles())
-                file.Delete();
-        }
-
-        public string GetIndexCheckpointFolder(Guid token = default(Guid))
-        {
-            if (token != default(Guid))
-                return GetMergedFolderPath(checkpointDir, index_base_folder, token.ToString());
-            else
-                return GetMergedFolderPath(checkpointDir, index_base_folder);
-        }
-
-        public string GetHybridLogCheckpointFolder(Guid token = default(Guid))
-        {
-            if (token != default(Guid))
-                return GetMergedFolderPath(checkpointDir, cpr_base_folder, token.ToString());
-            else
-                return GetMergedFolderPath(checkpointDir, cpr_base_folder);
-        }
-
-        public string GetIndexCheckpointMetaFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir,
-                                    index_base_folder,
-                                    token.ToString(),
-                                    index_meta_file,
-                                    ".dat");
-        }
-
-        public string GetPrimaryHashTableFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir,
-                                    index_base_folder,
-                                    token.ToString(),
-                                    hash_table_file,
-                                    ".dat");
-        }
-
-        public string GetOverflowBucketsFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir,
-                                    index_base_folder,
-                                    token.ToString(),
-                                    overflow_buckets_file,
-                                    ".dat");
-        }
-
-        public string GetHybridLogCheckpointMetaFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir,
-                                    cpr_base_folder,
-                                    token.ToString(),
-                                    cpr_meta_file,
-                                    ".dat");
-        }
-
-        public string GetHybridLogCheckpointContextFileName(Guid checkpointToken, Guid sessionToken)
-        {
-            return GetMergedFolderPath(checkpointDir,
-                                    cpr_base_folder,
-                                    checkpointToken.ToString(),
-                                    sessionToken.ToString(),
-                                    ".dat");
-        }
-
-        public string GetLogSnapshotFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir, cpr_base_folder, token.ToString(), snapshot_file, ".dat");
-        }
-
-        public string GetObjectLogSnapshotFileName(Guid token)
-        {
-            return GetMergedFolderPath(checkpointDir, cpr_base_folder, token.ToString(), snapshot_file, ".obj.dat");
-        }
-
-        private static string GetMergedFolderPath(params String[] paths)
-        {
-            String fullPath = paths[0];
-
-            for (int i = 1; i < paths.Length; i++)
-            {
-                if (i == paths.Length - 1 && paths[i].Contains("."))
-                {
-                    fullPath += paths[i];
-                }
-                else
-                {
-                    fullPath += Path.DirectorySeparatorChar + paths[i];
-                }
-            }
-
-            return fullPath;
         }
     }
 }
